refactored code and added yml env file
