+++ data_processing/dataset2df.py
@@ -1,8 +1,5 @@
-import numpy as np
-import matplotlib.pyplot as plt
 import os
 import json
-import cv2
 import pandas as pd
 
 def process_dataset_splitted(dataset_path, add_source = False):
