+++ data_processing/patch_ilm.py
@@ -1,5 +1,4 @@
 from tqdm import tqdm_notebook as tqdm
-import numpy as np
 import json
 import cv2
 import os
