+++ requirements.txt
@@ -1,9 +1,8 @@
-torchvision==0.12.0
-torch==1.9.0
-torchaudio==0.9.0
-torchsampler
-geffnet
-timm
-transformers
-vit_pytorch
-efficientnet_pytorch
+einops==0.6.0
+geffnet==1.0.2
+pit==0.5
+pyyaml==6.0
+transformers==4.10.2
+pytorch==1.9.0
+torchvision==0.10.0
+vit-pytorch==0.22.0
